Small fix for a compilation warning in RadiusLayer

+++ Packet++/src/RadiusLayer.cpp
@@ -68,7 +68,7 @@ RadiusAttribute RadiusLayer::addAttrAt(const RadiusAttributeBuilder& attrBuilder
 
 	if (!extendLayer(offset, sizeToExtend))
 	{
-		LOG_ERROR("Could not extend RadiusLayer in [%d] bytes", newAttr.getTotalSize());
+		LOG_ERROR("Could not extend RadiusLayer in [%d] bytes", (int)newAttr.getTotalSize());
 		return RadiusAttribute(NULL);
 	}
